Refactor InfoType and ApiResponse types into separate module
The `InfoType` type originally defined in the `contentModel.ts` has been moved to a new `apiModel.ts` file. This separation allows clearer distinction of types and better organization of code. References to `InfoType` in other files have been updated accordingly. Also, a new `ApiResponse` type is introduced in `apiModel.ts` to represent structured API response data. Changes in backend service have been made to return this new response type.
