+++ frontend/src/model/apiModel.ts
@@ -1,0 +1,8 @@
+export type InfoType = {
+  count: number | null;
+};
+
+export type ApiResponseType<T> = {
+  info: InfoType;
+  data: T;
+};
