+++ frontend/src/services/contentService.ts
@@ -1,5 +1,6 @@
 import axios from "axios";
-import { ContentType, InfoType } from "../model/contentModel.ts";
+import { ContentType } from "../model/contentModel.ts";
+import { InfoType } from "../model/apiModel.ts";
 
 function processSingleContent(rawContent: ContentType): ContentType {
   return {
